Extract DocumentDigestEntry
